+++ app/notifications/process_notifications.py
@@ -12,7 +12,7 @@ from app import redis_store
 from app.celery import provider_tasks
 from notifications_utils.clients import redis
 
-from app.celery import QueueNames
+from app.config import QueueNames
 from app.models import SMS_TYPE, Notification, KEY_TYPE_TEST, EMAIL_TYPE, ScheduledNotification
 from app.dao.notifications_dao import (dao_create_notification,
                                        dao_delete_notifications_and_history_by_id,
