+++ docker/Dockerfile
@@ -20,7 +20,6 @@ RUN \
 		zip \
 		libpq-dev \
 		jq \
-		libcurl4-openssl-dev \
 	&& echo "Clean up" \
 	&& rm -rf /var/lib/apt/lists/* /tmp/*
