+++ requirements.txt
@@ -13,7 +13,7 @@ flask-marshmallow==0.6.2
 Flask-Bcrypt==0.6.2
 credstash==1.8.0
 boto3==1.4.4
-celery==4.0.2
+celery==3.1.25
 monotonic==1.2
 statsd==3.2.1
 jsonschema==2.5.1
@@ -21,7 +21,6 @@ gunicorn==19.6.0
 docopt==0.6.2
 six==1.10.0
 iso8601==0.1.11
-pycurl==7.43.0
 
 # pin to minor version 3.1.x
 notifications-python-client>=3.1,<3.2
@@ -32,10 +31,4 @@ awscli-cwlogs>=1.4,<1.5
 
 git+https://github.com/alphagov/notifications-utils.git@17.5.3#egg=notifications-utils==17.5.3
 
-# Kombu is a library that celery uses under the hood.
-# Kombu v4.0.2 (which ships with celery v4.0.2) doesn't work with SQS due to problems with their use of boto2, so
-# Kombu migrated to boto3 - We're waiting for that to get a release version, and then to get a new version of celery
-# that pulls that in. Until that point, we should override the kombu version to get these SQS fixes.
-# Additionally, kombu master also includes a fix for the main process taking 100% CPU and not distributing tasks (!)
-# See https://github.com/celery/kombu/pull/693 and https://github.com/celery/kombu/pull/760
-https://github.com/celery/kombu/zipball/b2f21289284496efd89acea003ff9c24105b970e
+git+https://github.com/alphagov/boto.git@2.43.0-patch3#egg=boto==2.43.0-patch3
