+++ tests/app/service/test_send_one_off_notification.py
@@ -5,7 +5,7 @@ import pytest
 from notifications_utils.recipients import InvalidPhoneError
 
 from app.v2.errors import BadRequestError, TooManyRequestsError
-from app.celery import QueueNames
+from app.config import QueueNames
 from app.service.send_notification import send_one_off_notification
 from app.models import KEY_TYPE_NORMAL, PRIORITY, SMS_TYPE
